Add files via upload
This project was done by me and other two friends, Nyi Nyi Myo Zin Htet and Zin Zar Zar Lwin. This project is a senior project (final year project) for their bachelor study. In this project, I contributed in designing, building the bot body and in coding especially for sensors and PID control.